WIP Move relative and threshold to separate file
